+++ views/admin/index.php
@@ -10,10 +10,12 @@ use yii\helpers\Html;
 
     <div class="panel-body">
 
-        <?php if (!empty($version)): ?>
-            <div class="alert alert-success" role="alert"><?= Yii::t('OnlydocumentsModule.base', '<strong>DocumentServer</strong> successfully connected! - Installed version: {version}', ['version' => $version]); ?></div>
+        <?php if (!empty($view['version'])): ?>
+            <div class="alert alert-success" role="alert"><?= Yii::t('OnlydocumentsModule.base', '<strong>DocumentServer</strong> successfully connected! - Installed version: {version}', ['version' => $view['version']]); ?></div>
         <?php elseif (empty($model->serverUrl)): ?>
             <div class="alert alert-warning" role="alert"><?= Yii::t('OnlydocumentsModule.base', '<strong>DocumentServer</strong> not configured yet.'); ?></div>
+        <?php elseif (!empty($view['error']) && $view['error'] == 6): ?>
+            <div class="alert alert-danger" role="alert"><?= Yii::t('OnlydocumentsModule.base', '<strong>DocumentServer</strong> invalid JWT token.'); ?></div>
         <?php else: ?>
             <div class="alert alert-danger" role="alert"><?= Yii::t('OnlydocumentsModule.base', '<strong>DocumentServer</strong> not accessible.'); ?></div>
         <?php endif; ?>
@@ -21,6 +23,10 @@ use yii\helpers\Html;
         <?php $form = ActiveForm::begin(['id' => 'configure-form']); ?>
         <div class="form-group">
             <?= $form->field($model, 'serverUrl'); ?>
+        </div>
+
+        <div class="form-group">
+            <?= $form->field($model, 'jwtSecret'); ?>
         </div>
 
         <div class="form-group">